requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
instaloader==4.9.6
mysql-connector-python==8.0.32
numpy==1.24.2
protobuf==3.20.3
requests==2.28.2
urllib3==1.26.15
